requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
filelock==3.9.0
gdown==4.6.3
ibapi==10.20.1
idna==3.4
kaggle==1.5.12
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
pandas==1.5.3
Pillow==9.4.0
pip==22.3.1
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.0
pytz==2022.7.1
requests==2.28.2
setuptools==65.6.3
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
wheel==0.37.1
